+++ src/app/offer/provider-offer-card/provider-offer-card.component.ts
@@ -3,48 +3,56 @@ import { Offer, Service } from '../model/offer.model';
 import { Route, Router } from '@angular/router';
 import { OfferService } from '../offer.service';
 import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmDialogComponent } from '../../shared/confirm-dialog/confirm-dialog.component';
 @Component({
   selector: 'app-provider-offer-card',
   templateUrl: './provider-offer-card.component.html',
   styleUrl: './provider-offer-card.component.css'
 })
 export class ProviderOfferCardComponent {
-  @Output() deleteOfferEvent = new EventEmitter<number>();
+  @Output() deleteOfferEvent = new EventEmitter<void>();
   @Input() offer!: Service;
   constructor(
     private snackBar: MatSnackBar,
     private router: Router,
-    private offerService: OfferService
-  ) {}
+    private offerService: OfferService,
+    private dialog: MatDialog
+  ) { }
   editService(service: Service): void {
     this.offerService.setService(service);
     this.router.navigate(['/service-edit']);
   }
   onImageError(event: Event): void {
     (event.target as HTMLImageElement).src = 'https://via.placeholder.com/300x200.png?text=404+Not+Found';
-  }  
+  }
   deleteService(offerId: number): void {
-    const snackBarRef = this.snackBar.open(
-      'Are you sure you want to delete this offer?',
-      'Confirm',
-      {
-        duration: 5000,
-        verticalPosition: 'top', // Moves the snackbar to the top
-        horizontalPosition: 'center', // Optional, centers it horizontally
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      width: '400px',
+      data: { question: 'Are you sure you want to delete this offer?' }
+    });
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.offerService.deleteOffer(offerId).subscribe({
+          next: () => {
+            this.snackBar.open('Service deleted successfully!', 'Close', {
+              duration: 5000,
+              verticalPosition: 'top',
+            });
+            this.deleteOfferEvent.emit();
+          },
+          error: (err) => {
+            this.snackBar.open('Error when deleting the service, please try again.', 'Close', {
+              duration: 5000,
+              verticalPosition: 'top',
+            });
+          }
+        });
+      } else {
+        console.log('User cancelled the action');
       }
-    );
-  
-    snackBarRef.onAction().subscribe(() => {
-      this.offerService.deleteOffer(offerId).subscribe({
-        next: () => {
-          console.log('Offer deleted successfully.');
-          this.deleteOfferEvent.emit(offerId);
-        },
-        error: (err) => {
-          console.error('Error deleting offer:', err);
-        }
-      });
     });
   }
-  
+
 }
